<Refactor> Refactored pup_step.rb
